Refactor web app urlas
