+++ app/titles/flows/titles.js
@@ -1,16 +1,16 @@
 import { escapeMd } from '../../utils/escapeMd.js'
 import { GROUP_CHAT_TYPES } from '../../shared/middlewares/groupChat.js'
 
-export function titleCommand({ generateWebAppUrl }) {
+export function titlesCommand({ generateWebAppUrl }) {
   return async (context) => {
     const { chatId, localize } = context.state
 
     const isGroup = GROUP_CHAT_TYPES.includes(context.chat.type)
 
-    const editUrl = isGroup ? generateWebAppUrl('group', chatId, 'titles') : generateWebAppUrl('groups')
+    const editUrl = isGroup ? generateWebAppUrl(`titles-${chatId}`) : generateWebAppUrl('groups')
 
     await context.reply(
-      localize('command.title.help', {
+      localize('command.titles.help', {
         editUrl: escapeMd(editUrl),
       }),
       { parse_mode: 'MarkdownV2', disable_web_page_preview: true }
