+++ app/utils/createWebAppUrlGenerator.js
@@ -1,6 +1,6 @@
 export function createWebAppUrlGenerator({ botUsername, webAppName }) {
-  return (...commands) => {
-    const query = commands.length > 0 ? `?startapp=${commands.join('__')}` : ''
+  return (command) => {
+    const query = command ? `?startapp=${command}` : ''
     return `https://t.me/${botUsername}/${webAppName}${query}`
   }
 }
