+++ assets/localization/uk.json
@@ -80,7 +80,7 @@
         "{viewUrl}"
       ]
     },
-    "title": {
+    "titles": {
       "help": [
         "✍️ Відредагувати **підписи** можна тут:",
         "{editUrl}"
