+++ package.json
@@ -54,6 +54,7 @@
     "markdown-escapes": "^2.0.1",
     "multer": "^1.4.4",
     "nanoid": "^5.0.5",
+    "nanoid-dictionary": "^5.0.0-beta.1",
     "pg": "^8.7.1",
     "pino": "^7.9.1",
     "pino-pretty": "^7.5.4",
